+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.5] - 2016-02-26
+### Added
+- Added `Helper` class that provides a static method to parse parameters.
+
+### Fixed
+- Refactored traits to deduplicate code.
+
 ## [0.1.4] - 2016-02-26
 ### Fixed
 - Removed unused local variables in tests.
@@ -22,6 +29,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release to GitHub.
 
+[0.1.5]: https://github.com/brightnucleus/invoker/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/brightnucleus/invoker/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/brightnucleus/invoker/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/brightnucleus/invoker/compare/v0.1.1...v0.1.2